cloudfoundry-worker/src/main/java/io/pivotal/conductor/worker/cloudfoundry/DeriveCloudFoundryResourceNameWorker.java
package io.pivotal.conductor.worker.cloudfoundry;

import com.netflix.conductor.client.worker.Worker;
import com.netflix.conductor.common.metadata.tasks.Task;
import com.netflix.conductor.common.metadata.tasks.TaskResult;
import com.netflix.conductor.common.metadata.tasks.TaskResult.Status;

public class DeriveCloudFoundryResourceNameWorker implements Worker {

    static final String TASK_DEF_NAME = "derive_cloud_foundry_resource_name";

    @Override
    public String getTaskDefName() {
        return TASK_DEF_NAME;
    }

    @Override
    public TaskResult execute(Task task) {
        String projectName = (String) task.getInputData().get("projectName");

        TaskResult taskResult = new TaskResult(task);
        taskResult.setStatus(Status.COMPLETED);
        taskResult.addOutputData("resourceName", CloudFoundryUtil.deriveResourceName(projectName));
        return taskResult;
    }
}
